+++ esbuild-watch.js
@@ -1,16 +1,25 @@
-const executeBuildCommand = require('./executeBuildCommand');
+const esbuild = require('esbuild');
+const { esmConfig } = require('./esbuild.config');
 
-executeBuildCommand()({
-  watch: {
-    onRebuild(error, result) {
-      if (error) {
-        console.error('Watch build failed:', error);
-      } else {
-        console.log(
-          '\x1b[36m%s\x1b[0m',
-          `[${new Date().toLocaleTimeString()}] sdk-dapp-form rebuild succeeded`
-        );
+esbuild
+  .build({
+    ...esmConfig,
+    minify: false,
+    define: {
+      ...esmConfig.define,
+      'process.env.NODE_ENV': '"development"'
+    },
+    watch: {
+      onRebuild(error) {
+        if (error) {
+          console.error('Watch build failed:', error);
+        } else {
+          console.log(
+            '\x1b[36m%s\x1b[0m',
+            `[${new Date().toLocaleTimeString()}] sdk-dapp-form rebuild succeeded`
+          );
+        }
       }
     }
-  }
-});
+  })
+  .catch(() => process.exit(1));